Fix miscellanious size_t related compiler warnings.

+++ contrib/ipfilter/ipft_ef.c
@@ -146,7 +146,7 @@ int	cnt, *dir;
 	(void) inet_aton(src, &ip->ip_src);
 	(void) inet_aton(dst, &ip->ip_dst);
 	ip->ip_len = atoi(len);
-	ip->ip_hl = sizeof(ip_t);
+	ip->ip_hl = (unsigned)sizeof(ip_t);
 
 	slen = ip->ip_hl + extra;
 	i = MIN(cnt, slen);
